:recycle: Refactor: AppleUtils 클래스 분리
<footer>
- 관련: #134
