+++ application/external-api/src/main/java/com/namo/spring/application/external/domain/user/application/impl/UserService.java
@@ -1,14 +1,8 @@
 package com.namo.spring.application.external.domain.user.application.impl;
 
-import java.io.IOException;
-import java.io.Reader;
-import java.io.StringReader;
 import java.math.BigInteger;
-import java.nio.file.Files;
-import java.nio.file.Paths;
 import java.security.KeyFactory;
 import java.security.NoSuchAlgorithmException;
-import java.security.PrivateKey;
 import java.security.PublicKey;
 import java.security.spec.InvalidKeySpecException;
 import java.security.spec.RSAPublicKeySpec;
@@ -18,13 +12,9 @@ import java.util.Date;
 import java.util.List;
 import java.util.Optional;
 
-import org.bouncycastle.asn1.pkcs.PrivateKeyInfo;
-import org.bouncycastle.openssl.PEMParser;
-import org.bouncycastle.openssl.jcajce.JcaPEMKeyConverter;
 import org.json.simple.JSONObject;
 import org.json.simple.parser.JSONParser;
 import org.json.simple.parser.ParseException;
-import org.springframework.core.io.ClassPathResource;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Service;
 import org.springframework.util.StringUtils;
@@ -36,9 +26,9 @@ import com.namo.spring.application.external.domain.user.domain.constant.UserStat
 import com.namo.spring.application.external.domain.user.repository.TermRepository;
 import com.namo.spring.application.external.domain.user.repository.UserRepository;
 import com.namo.spring.application.external.domain.user.ui.dto.UserRequest;
-import com.namo.spring.application.external.global.feignclient.apple.AppleProperties;
-import com.namo.spring.application.external.global.feignclient.apple.AppleResponse;
 import com.namo.spring.application.external.global.utils.JwtUtils;
+import com.namo.spring.client.social.apple.AppleProperties;
+import com.namo.spring.client.social.apple.AppleResponse;
 import com.namo.spring.core.common.code.status.ErrorStatus;
 import com.namo.spring.core.common.exception.UserException;
 
@@ -189,20 +179,5 @@ public class UserService {
 			throw new IllegalArgumentException("Token expired");
 		}
 		return true;
-	}
-
-	public PrivateKey getPrivateKey() {
-		try {
-			ClassPathResource resource = new ClassPathResource(appleProperties.getPrivateKeyPath());
-			String privateKey = new String(Files.readAllBytes(Paths.get(resource.getURI())));
-			Reader pemReader = new StringReader(privateKey);
-
-			PEMParser pemParser = new PEMParser(pemReader);
-			JcaPEMKeyConverter converter = new JcaPEMKeyConverter();
-			PrivateKeyInfo object = (PrivateKeyInfo)pemParser.readObject();
-			return converter.getPrivateKey(object);
-		} catch (IOException e) {
-			throw new RuntimeException(e);
-		}
 	}
 }
